add: protect user paths using roles

+++ src/users/users.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Body, Param, Delete, Patch, Get, Post, Req } from '@nestjs/common';
+import { Controller, Body, Param, Delete, Patch, Get, Post, Req, ForbiddenException } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -6,13 +6,17 @@ import { buildResponse } from '../common/utils/build-response';
 import { ValidateObjectIdPipe } from '../common/pipes/validate-object-id.pipe';
 import { UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
+import { RolesGuard } from './guards/role.guard';
+import { Roles } from './decorators/roles.decorator';
 
-@UseGuards(AuthGuard('jwt'))
+
+@UseGuards(AuthGuard('jwt'), RolesGuard)
 @Controller('users')
 export class UsersController {
     constructor(private readonly usersService: UsersService) {}
 
     @Get()
+    @Roles('admin')
     async findAll(@Req() req) {
         const users = await this.usersService.findAll();
 
@@ -25,8 +29,15 @@ export class UsersController {
 
     @Get(':id')
     async findById(@Param('id', ValidateObjectIdPipe) id: string, @Req() req) {
-        const user = await this.usersService.findById(id);
-        return buildResponse(true, 'User retrieved successfully', user, req);
+        const { user } = req;
+
+        console.log(user);
+        if (user.role !== 'admin' && user.id !== id) {
+            throw new ForbiddenException('You can only access your own profile');
+        }
+
+        const foundUser = await this.usersService.findById(id);
+        return buildResponse(true, 'User retrieved successfully', foundUser, req);
     }
 
     @Post()
@@ -37,13 +48,23 @@ export class UsersController {
 
     @Patch(':id')
     async updateUser(@Param('id', ValidateObjectIdPipe) id: string, @Body() dto: UpdateUserDto, @Req() req) {
-        const user = await this.usersService.update(id, dto);
-        return buildResponse(true, 'User updated successfully', user, req);
+        const {user} = req;
+        if (user.role !== 'admin' && user.id !== id) {
+            throw new ForbiddenException('You can only update your own profile');
+        }
+
+        const foundUser = await this.usersService.update(id, dto);
+        return buildResponse(true, 'User updated successfully', foundUser, req);
     }
 
     @Delete(':id')
     async removeUser(@Param('id', ValidateObjectIdPipe) id: string, @Req() req) {
-        const user = await this.usersService.remove(id);
-        return buildResponse(true, 'User deleted successfully', user, req);
+        const {user} = req;
+        if (user.role !== 'admin' && user.id !== id) {
+            throw new ForbiddenException('You can only delete your own profile');
+        }
+
+        const foundUser = await this.usersService.remove(id);
+        return buildResponse(true, 'User deleted successfully', foundUser, req);
     }
 }
